XBee Dynamic Storage
This sketch will cause an XBee radio attached to the Arduino to listen
for requests from another XBee radio. It will take the
payload received by the XBee and add it to a dynamically allocated
array of strings (character arrays). It will attempt to grow the array
holding the strings as it adds them.